fix(core): keep bigint values lossless at runtime and add bigintMode (#481)
Drivers rounded BIGINT beyond 2^53 before the ORM saw it (mysql2 without supportBigNumbers, better-sqlite3 without safeIntegers) while pg returned strings. Connectors now deliver integers losslessly and the new column option bigintMode ("number" default with BIGINT_PRECISION_LOSS instead of rounding, "string", "bigint") picks one entity type on every driver; aggregates, batch PK derivation, cursors and logging follow.

+++ __tests__/unit/golden-sql/dialect-primitives.golden.test.ts
@@ -26,7 +26,7 @@ describe("golden-sql / castBuiltinType — ColumnType to dialect SQL type", () =
     ["number", "INTEGER", "INT", "INTEGER"],
     ["float", "REAL", "FLOAT", "REAL"],
     ["double", "NUMERIC($precision, $scale)", "DECIMAL($precision, $scale)", "REAL"],
-    ["bigint", "BIGINT", "BIGINT", "INTEGER"],
+    ["bigint", "BIGINT", "BIGINT", "BIGINT"],
     ["boolean", "BOOLEAN", "TINYINT($n)", "INTEGER"],
     ["datetime", "TIMESTAMP", "DATETIME", "TEXT"],
     ["timestamp", "TIMESTAMP", "TIMESTAMP", "TEXT"],
